+++ RegularExpression/OGMutableAttributedString.h
@@ -20,8 +20,8 @@
 {
 	NSString		*_currentFontFamilyName;
 	NSFontTraitMask	_currentFontTraits;
-	float			_currentFontWeight;
-	float			_currentFontPointSize;
+	CGFloat			_currentFontWeight;
+	CGFloat			_currentFontPointSize;
 	NSDictionary	*_currentAttributes;
 	NSFontManager	*_fontManager;
 }
